Vax dev to master

+++ Layers/xrRender/DetailManager_CACHE.cpp
@@ -17,9 +17,9 @@ void CDetailManager::cache_Initialize()
 		}
 	VERIFY(cache_Validate());
 
-	for (int _mz1 = 0; _mz1 < dm_cache1_line; _mz1++)
+	for (u32 _mz1 = 0; _mz1 < dm_cache1_line; _mz1++)
 	{
-		for (int _mx1 = 0; _mx1 < dm_cache1_line; _mx1++)
+		for (u32 _mx1 = 0; _mx1 < dm_cache1_line; _mx1++)
 		{
 			CacheSlot1 &MS = cache_level1[_mz1][_mx1];
 			for (int _z = 0; _z < dm_cache1_count; _z++)
@@ -32,9 +32,9 @@ void CDetailManager::cache_Initialize()
 CDetailManager::Slot *CDetailManager::cache_Query(int r_x, int r_z)
 {
 	int gx = w2cg_X(r_x + cache_cx);
-	VERIFY(gx >= 0 && gx < dm_cache_line);
+	VERIFY(gx >= 0 && (u32)gx < dm_cache_line);
 	int gz = w2cg_Z(r_z + cache_cz);
-	VERIFY(gz >= 0 && gz < dm_cache_line);
+	VERIFY(gz >= 0 && (u32)gz < dm_cache_line);
 	return cache[gz][gx];
 }
 
@@ -76,9 +76,9 @@ void CDetailManager::cache_Task(int gx, int gz, Slot *D)
 
 BOOL CDetailManager::cache_Validate()
 {
-	for (int z = 0; z < dm_cache_line; z++)
+	for (u32 z = 0; z < dm_cache_line; z++)
 	{
-		for (int x = 0; x < dm_cache_line; x++)
+		for (u32 x = 0; x < dm_cache_line; x++)
 		{
 			int w_x = cg2w_X(x);
 			int w_z = cg2w_Z(z);
@@ -103,10 +103,10 @@ void CDetailManager::cache_Update(int v_x, int v_z, Fvector &view, int limit)
 		{
 			// shift matrix to left
 			cache_cx++;
-			for (int z = 0; z < dm_cache_line; z++)
+			for (u32 z = 0; z < dm_cache_line; z++)
 			{
 				Slot *S = cache[z][0];
-				for (int x = 1; x < dm_cache_line; x++)
+				for (u32 x = 1; x < dm_cache_line; x++)
 					cache[z][x - 1] = cache[z][x];
 				cache[z][dm_cache_line - 1] = S;
 				cache_Task(dm_cache_line - 1, z, S);
@@ -117,10 +117,10 @@ void CDetailManager::cache_Update(int v_x, int v_z, Fvector &view, int limit)
 		{
 			// shift matrix to right
 			cache_cx--;
-			for (int z = 0; z < dm_cache_line; z++)
+			for (u32 z = 0; z < dm_cache_line; z++)
 			{
 				Slot *S = cache[z][dm_cache_line - 1];
-				for (int x = dm_cache_line - 1; x > 0; x--)
+				for (u32 x = dm_cache_line - 1; x > 0; x--)
 					cache[z][x] = cache[z][x - 1];
 				cache[z][0] = S;
 				cache_Task(0, z, S);
@@ -134,7 +134,7 @@ void CDetailManager::cache_Update(int v_x, int v_z, Fvector &view, int limit)
 		{
 			// shift matrix down a bit
 			cache_cz++;
-			for (int x = 0; x < dm_cache_line; x++)
+			for (u32 x = 0; x < dm_cache_line; x++)
 			{
 				Slot *S = cache[dm_cache_line - 1][x];
 				for (int z = dm_cache_line - 1; z > 0; z--)
@@ -148,10 +148,10 @@ void CDetailManager::cache_Update(int v_x, int v_z, Fvector &view, int limit)
 		{
 			// shift matrix up
 			cache_cz--;
-			for (int x = 0; x < dm_cache_line; x++)
+			for (u32 x = 0; x < dm_cache_line; x++)
 			{
 				Slot *S = cache[0][x];
-				for (int z = 1; z < dm_cache_line; z++)
+				for (u32 z = 1; z < dm_cache_line; z++)
 					cache[z - 1][x] = cache[z][x];
 				cache[dm_cache_line - 1][x] = S;
 				cache_Task(x, dm_cache_line - 1, S);
@@ -206,9 +206,9 @@ void CDetailManager::cache_Update(int v_x, int v_z, Fvector &view, int limit)
 
 	if (bNeedMegaUpdate)
 	{
-		for (int _mz1 = 0; _mz1 < dm_cache1_line; _mz1++)
+		for (u32 _mz1 = 0; _mz1 < dm_cache1_line; _mz1++)
 		{
-			for (int _mx1 = 0; _mx1 < dm_cache1_line; _mx1++)
+			for (u32 _mx1 = 0; _mx1 < dm_cache1_line; _mx1++)
 			{
 				CacheSlot1 &MS = cache_level1[_mz1][_mx1];
 				MS.empty = TRUE;
